requirements.txt
numpy==1.21.*
scipy==1.7.3
matplotlib==3.5.1
pandas==1.4.1
scikit-learn==1.0.2
scikit-image==0.19.2
PyYAML==5.1.*
sklearn-pandas==1.8.0
sktime==0.12.0
hydra-core==1.1.1
optuna==2.10.0
hydra-optuna-sweeper==1.1.1
mlflow==1.26.1
mlflow_extend
plotly==5.8.0
catboost
